test(pornhub): add failing tests for HTML search parser
