+++ crates/rdlp-extractor/src/extractors/pornhub/tests/mod.rs
@@ -1,0 +1,7 @@
+//! Tests for the PornHub extractor.
+
+#[cfg(test)]
+mod extractor;
+
+#[cfg(test)]
+mod search_html;
